+++ src/main/scala/com/exasol/cloudetl/bucket/GCSBucket.scala
@@ -16,6 +16,10 @@ final case class GCSBucket(path: String, params: StorageProperties) extends Buck
   /** @inheritdoc */
   override val properties: StorageProperties = params
 
+  /** @inheritdoc */
+  override def validate(): Unit =
+    validateRequiredProperties()
+
   /**
    * Returns the list of required property keys for Google Cloud
    * Storage.