+++ src/test/scala/com/exasol/cloudetl/bucket/AbstractBucketTest.scala
@@ -1,9 +1,16 @@
 package com.exasol.cloudetl.bucket
 
+import com.exasol.ExaConnectionInformation
+import com.exasol.ExaMetadata
+import com.exasol.cloudetl.storage.StorageProperties
+
+import org.mockito.Mockito.when
 import org.scalatest.BeforeAndAfterEach
 import org.scalatest.FunSuite
+import org.scalatest.mockito.MockitoSugar
 
-class AbstractBucketTest extends FunSuite with BeforeAndAfterEach {
+@SuppressWarnings(Array("org.wartremover.warts.Overloading"))
+class AbstractBucketTest extends FunSuite with BeforeAndAfterEach with MockitoSugar {
 
   private[bucket] val PATH: String = "BUCKET_PATH"
   private[bucket] val FORMAT: String = "DATA_FORMAT"
@@ -14,4 +21,28 @@ class AbstractBucketTest extends FunSuite with BeforeAndAfterEach {
     ()
   }
 
+  protected[this] final def getBucket(params: Map[String, String]): Bucket =
+    Bucket(StorageProperties(params))
+
+  protected[this] final def getBucket(
+    params: Map[String, String],
+    exaMetadata: ExaMetadata
+  ): Bucket =
+    Bucket(StorageProperties(params, exaMetadata))
+
+  protected[this] final def mockConnectionInfo(
+    username: String,
+    password: String
+  ): ExaMetadata = {
+    val metadata = mock[ExaMetadata]
+    val connectionInfo: ExaConnectionInformation = new ExaConnectionInformation() {
+      override def getType(): ExaConnectionInformation.ConnectionType =
+        ExaConnectionInformation.ConnectionType.PASSWORD
+      override def getAddress(): String = ""
+      override def getUser(): String = username
+      override def getPassword(): String = password
+    }
+    when(metadata.getConnection("connection_info")).thenReturn(connectionInfo)
+    metadata
+  }
 }